Increase the period works? wait before unblocking the main thread.
There are two threads running when we are testing if an adapter works or not:
the main thread and the "guarding" thread. The guarding thread only ensures
that errors won't result in an indefinite blockage. Increasing the waiting time
would give the main thread more time which could be usefull on slow machines.