Fixes the #silenced? method when "only" patterns are present
As soon as a path is matched against an "only" regexp, it's
not silenced, so it's silenced if it doesn't match any "only"
regexps.
If fixes the previous logic that was requiring a path to match
all "only" patterns in order to not be silenced, meaning that
it was silenced as soon as it didn't match an "only" regexp
(even if it matched another "only" regexp).