Use proper attempts variable on ECONNREFUSED
Previous to this commit, the Listen::Adapter#start method impcorrectly
used the 'retries' variable instead of the correct 'attempts' variable to
count the number of attempts that have occurred.

Further, the attempts variable was reset to the value of 3 on every
retry causing indefinite retries.

This commit renames the 'retries' variable to 'attempts' and also only
sets the 'attempts' variable to a value of 3 if it does not already have
a value.