Don't require rb-fsevent until actually needed
rb-fsevent is used only on macOS. If you somehow exclude it from
dependencies on other platforms, for example to avoid unnecessary
complications with building native extensions, then "listen" fails to
start just because it's required before checked that it's not usable on
the current platform. This patch fixes this error.

See also #416