Avoid warnings for calls to ordered on RSpec stubs
It's only possible to constrain the order in which calls are made when
using `expect`; not when using `allow`. Indeed you need at least two
calls to `expect` with the `ordered` qualifier to make this work.

The following warning was appearing in the spec output multiple times:

    WARNING: `allow(...).to receive(..).ordered` is not supported and will have no effect, use `and_return(*ordered_values)` instead..

Thus the calls to `ordered` were not having any effect and so I have
removed them.

See the documentation [1] for more info.

[1]: https://relishapp.com/rspec/rspec-mocks/v/3-8/docs/setting-constraints/message-order