Avoid warnings due to multiple calls to configure
I was seeing multiple warnings like the following in the spec output:

    W, [2019-08-07T11:42:43.494699 #48484]  WARN -- : Adapter already configured!

This was because in the Adapter::Darwin spec in the "running threads"
describe block, both `subject.configure` and `subject.start` were being
called and the latter calls the former.

By moving the call to `subject.configure` into the "configuration"
describe block, we can avoid this problem.