Bump required ruby version to v2.4.0
a9137acc introduced a use of Hash#transform_values in
Listen::Record#dir_entries, however that was only added to Ruby in
[v2.4](https://github.com/ruby/ruby/blob/v2_4_0/NEWS#core-classes-updates-outstanding-ones-only-).

This can lead to runtime errors when run on an older version of Ruby,
such as 2.2.7, which the gemspec previously implied was supported.

Also expand CI to match and test on Ruby v2.4.