Raise argument error when passing a file path
The Listen gem doesn't support listening on changes of individual files,
but currently if we pass a file path to `Listen.on`, the listener is
successfully initialized, without actually listening on given files. To
improve clarity, we notify the user they need to pass a directory by
raising an argument error.