Use the defined type to the default value of `directory`
Currently, you will get Thor's deprecated message when starting the `Listen::CLI`.

```
Deprecation warning: Expected array default value for '--directory'; got "." (string).
This will be rejected in the future unless you explicitly pass the options `check_default_type: false` or call `allow_incompatible_default_type!` in your code
You can silence deprecations warning by setting the environment variable THOR_SILENCE_DEPRECATION.
```

This is due to the incorrect default value(`directory` is defined as an `array`,
but the default value is a `string`).

This fixed to use the correct default value and correctly pass to the `directory`
to `Listen.to`.